Refactor AddressSpaceFilter component, create container component and create separate kabab component
