+++ console/console-init/ui/src/modules/address-space/components/AddressSpaceListKebab/AddressSpaceListKebab.tsx
@@ -1,0 +1,86 @@
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
+import React from "react";
+import {
+  OverflowMenu,
+  OverflowMenuContent,
+  OverflowMenuGroup,
+  OverflowMenuItem,
+  OverflowMenuControl
+} from "@patternfly/react-core/dist/js/experimental";
+import {
+  Dropdown,
+  DropdownItem,
+  Button,
+  ButtonVariant,
+  KebabToggle
+} from "@patternfly/react-core";
+
+export interface IAddressSpaceListKebabProps {
+  onCreateAddressSpace: () => void;
+  isDeleteAllDisabled: boolean;
+  onKebabSelect: (event: any) => void;
+}
+
+export const AddressSpaceListKebab: React.FC<IAddressSpaceListKebabProps> = ({
+  isDeleteAllDisabled,
+  onCreateAddressSpace,
+  onKebabSelect
+}) => {
+  const [isKebabOpen, setIsKebabOpen] = React.useState(false);
+
+  const onKebabToggle = (isOpen: boolean) => {
+    setIsKebabOpen(isOpen);
+  };
+
+  const dropdownItems = [
+    <DropdownItem
+      id="as-list-delete-all"
+      key="delete-all"
+      component="button"
+      value="deleteAll"
+      isDisabled={isDeleteAllDisabled}
+    >
+      Delete Selected
+    </DropdownItem>
+  ];
+
+  return (
+    <>
+      <OverflowMenu breakpoint="lg">
+        <OverflowMenuContent isPersistent>
+          <OverflowMenuGroup groupType="button" isPersistent>
+            {/* Remove is Persistent after fixing dropdown items for overflow menu */}
+            <OverflowMenuItem isPersistent>
+              <Button
+                id="al-filter-overflow-button"
+                variant={ButtonVariant.primary}
+                onClick={onCreateAddressSpace}
+              >
+                Create Address Space
+              </Button>
+            </OverflowMenuItem>
+          </OverflowMenuGroup>
+        </OverflowMenuContent>
+        <OverflowMenuControl hasAdditionalOptions>
+          <Dropdown
+            id="al-filter-overflow-dropdown"
+            onSelect={onKebabSelect}
+            toggle={
+              <KebabToggle
+                id="al-filter-overflow-kebab"
+                onToggle={onKebabToggle}
+              />
+            }
+            isOpen={isKebabOpen}
+            isPlain
+            dropdownItems={dropdownItems}
+          />
+        </OverflowMenuControl>
+      </OverflowMenu>
+    </>
+  );
+};
